Correct the condition used to terminate the compress / decompress loop
Previously the loop terminated when the output buffer was not completely filled by the previous call. Most of the time this is correct, because if the output buffer was completely filled, there is probably more to come. However, in the situation that the output fits perfectly in the output buffer, the buffer will be completely filled, yet there is nothing left to process.  In this case we would run through the loop again calling inflate() / deflate() which would return Z_BUF_ERROR. This would manifest itself as the request (needlessly) failing.

Rather than this we should simply terminate the loop when there is no more input left to process.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# Committer: Alex Deem <alexdeem@Alex-MacBook-Pro.local>
#
# On branch master
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   Classes/ASIDataCompressor.m
#	modified:   Classes/ASIDataDecompressor.m
#